api delegator
